DL: Refactor gp_segment_id_col
This was getting passed around as a param to a lot of different
functions, but all it was really doing is serving as a flag
to say whether we're running on greenplum or postgres.  Now this
is just determined right before the query by calling is_platform_pg().

Other minor chages:
  - Removed extraneous error check in fit_transiton
  - Renamed rows_per_seg to images_per_seg param in compute_metrics.
  - Remove STRICT property from internal_keras_eval_transition()
       and internal_keras_eval_final().  We think this may result
       in slightly better error detection.

Co-authored-by: Orhan Kislal <okislal@apache.org>